add custom config

+++ run.sh
@@ -5,49 +5,32 @@ set -o nounset
 set -o pipefail
 
 ROOT_DIR=$(dirname "${BASH_SOURCE[0]}")
-KARMADA_REPO=${KARMADA_REPO:-"${ROOT_DIR}/../karmada"}
-source "${KARMADA_REPO}/hack/util.sh"
 source "${ROOT_DIR}/util.sh"
 
+source "${ROOT_DIR}/default.config"
+[[ -f "${ROOT_DIR}/config" ]] && source "${ROOT_DIR}/config"
+
+CERT_DIR=${KARMADA_DIR}
+ROOT_CA_FILE=${CERT_DIR}/server-ca.crt
+LAUNCH_DIR="${HOME}/Library/LaunchAgents"
 CFSSL_VERSION="v1.5.0"
 
-KARMADA_DIR="${HOME}/.karmada"
-BIN_DIR="${HOME}/bin"
-LOG_DIR="${KARMADA_DIR}/logs"
-CERT_DIR=${KARMADA_DIR}
-ETCD_DIR=${KARMADA_DIR}/etcd
-LAUNCH_DIR="${HOME}/Library/LaunchAgents"
-
-ROOT_CA_FILE=${CERT_DIR}/server-ca.crt
-
-KARMADA_KUBECONFIG=${KARMADA_KUBECONFIG:-"${KARMADA_DIR}/karmada-apiserver.config"}
 KARMADA_APISERVER_IP=127.0.0.1
-KARMADA_APISERVER_SECURE_PORT=5443
 
 KUBECTL="kubectl --kubeconfig=${KARMADA_KUBECONFIG} --context karmada-apiserver"
 
 install() {
   echo ! Starting install karmada
-
-  echo install karmada with:
-  "${BIN_DIR}"/etcd --version | grep etcd
-  echo "kube-apiserver version::: $("${BIN_DIR}"/kube-apiserver --version)"
-  echo "kube-controller-manage version::r: $("${BIN_DIR}"/kube-controller-manager --version)"
-  "${BIN_DIR}"/karmada-aggregated-apiserver version
-  "${BIN_DIR}"/karmada-controller-manager version
-  "${BIN_DIR}"/karmada-scheduler version
-  "${BIN_DIR}"/karmada-descheduler version
-  "${BIN_DIR}"/karmada-webhook version
-  "${BIN_DIR}"/karmada-search version
-#  "${BIN_DIR}"/karmada-scheduler-estimator version
+  print_binaries_version
 
   gen_cert
   install_launch_tasks
   start
 
-  util:wait_until ${KUBECTL} get ns > /dev/null
-
+  util::wait_until ${KUBECTL} version -o yaml > /dev/null
   install_kube_artifacts
+
+  print_success
 }
 
 uninstall() {
@@ -72,59 +55,65 @@ stop() {
 }
 
 status() {
+  echo 'NOTE: It is reported by `launchctl list`. See more about it by `man launchctl`'
   launchctl list | grep -E "PID|com.github.karmada-io.*"
 }
 
 help() {
   cat <<EOF
-Usage: $0 [start|stop|status|clean|help]
+Usage: $0 <COMMAND>
+
+Commands:
+install   : install config files, and start karmada.
+uninstall : stop karmada, and remove config files, and clean data.
+start     : run the karmada.
+stop      : stop the karmada.
+status    : show the status of processes.
+help      : print the usage of this script.
 EOF
 }
 
 gen_cert() {
-  if [[ -f "${CERT_DIR}"/tls.crt ]]; then
-    echo !!! cert existed, skip.
-    return
+  if [[ ! -f "${CERT_DIR}"/front-proxy-ca-config.json ]]; then
+      echo !!! generate cert
+      util::cmd_must_exist "openssl"
+      util::cmd_must_exist_cfssl ${CFSSL_VERSION}
+
+      # create CA signers
+      rm -rf "${CERT_DIR}"
+      mkdir -p "${CERT_DIR}"
+      util::create_signing_certkey "" "${CERT_DIR}" server '"client auth","server auth"'
+      util::create_signing_certkey "" "${CERT_DIR}" front-proxy '"client auth","server auth"'
+
+      # signs a certificate
+      util::create_certkey "" "${CERT_DIR}" "server-ca" \
+    	karmada system:admin kubernetes.default.svc \
+    	"*.etcd.karmada-system.svc.cluster.local" \
+    	"*.karmada-system.svc.cluster.local" \
+    	"*.karmada-system.svc" \
+    	"localhost" \
+    	"127.0.0.1"
+
+      util::create_certkey "" "${CERT_DIR}" "front-proxy-ca" \
+    	front-proxy-client \
+    	front-proxy-client kubernetes.default.svc \
+    	"*.etcd.karmada-system.svc.cluster.local" \
+    	"*.karmada-system.svc.cluster.local" \
+    	"*.karmada-system.svc" \
+    	"localhost" \
+    	"127.0.0.1"
   fi
 
-  echo !!! generate cert
-  util::cmd_must_exist "openssl"
-  util::cmd_must_exist_cfssl ${CFSSL_VERSION}
-
-  # create CA signers
-  rm -rf "${CERT_DIR}"
-  mkdir -p "${CERT_DIR}"
-  util::create_signing_certkey "" "${CERT_DIR}" server '"client auth","server auth"'
-  util::create_signing_certkey "" "${CERT_DIR}" front-proxy '"client auth","server auth"'
-
-  # signs a certificate
-  util::create_certkey "" "${CERT_DIR}" "server-ca" \
-	karmada system:admin kubernetes.default.svc \
-	"*.etcd.karmada-system.svc.cluster.local" \
-	"*.karmada-system.svc.cluster.local" \
-	"*.karmada-system.svc" \
-	"localhost" \
-	"127.0.0.1"
-
-  util::create_certkey "" "${CERT_DIR}" "front-proxy-ca" \
-	front-proxy-client \
-	front-proxy-client kubernetes.default.svc \
-	"*.etcd.karmada-system.svc.cluster.local" \
-	"*.karmada-system.svc.cluster.local" \
-	"*.karmada-system.svc" \
-	"localhost" \
-	"127.0.0.1"
-
   # tls for webhook
-  ln -s "${CERT_DIR}"/server-ca.key "${CERT_DIR}"/tls.key
-  ln -s "${CERT_DIR}"/server-ca.crt "${CERT_DIR}"/tls.crt
+  [[ ! -f "${CERT_DIR}"/tls.key ]] && ln -s "${CERT_DIR}"/server-ca.key "${CERT_DIR}"/tls.key
+  [[ ! -f "${CERT_DIR}"/tls.crt ]] && ln -s "${CERT_DIR}"/server-ca.crt "${CERT_DIR}"/tls.crt
 
   # write karmada api server config to kubeconfig file
   util::append_client_kubeconfig "${KARMADA_KUBECONFIG}" \
 	"${CERT_DIR}/karmada.crt" "${CERT_DIR}/karmada.key" \
 	"${KARMADA_APISERVER_IP}" "${KARMADA_APISERVER_SECURE_PORT}" \
 	karmada-apiserver
-  "${KUBECTL}" config use-context karmada-apiserver
+  ${KUBECTL} config use-context karmada-apiserver
 }
 
 install_launch_tasks() {
@@ -133,7 +122,21 @@ install_launch_tasks() {
     for file in *.plist; do
       if [[ ! -f "${LAUNCH_DIR}/${file}" ]]; then
         echo !!! install launch task: "${file}"
-        sed "s|{{HOME}}|${HOME}|g" < "${file}" > "${LAUNCH_DIR}/${file}"
+        # shellcheck disable=SC2002
+        cat "${file}" \
+            | sed "s|{{BIN_DIR}}|${BIN_DIR}|g" \
+            | sed "s|{{LOG_DIR}}|${LOG_DIR}|g" \
+            | sed "s|{{KARMADA_DIR}}|${KARMADA_DIR}|g" \
+            | sed "s|{{KARMADA_KUBECONFIG}}|${KARMADA_KUBECONFIG}|g" \
+            | sed "s|{{ETCD_PORT}}|${ETCD_PORT}|g" \
+            | sed "s|{{ETCD_PEER_PORT}}|${ETCD_PEER_PORT}|g" \
+            | sed "s|{{KARMADA_APISERVER_SECURE_PORT}}|${KARMADA_APISERVER_SECURE_PORT}|g" \
+            | sed "s|{{KARMADA_AGGREGATED_APISERVER_SECURE_PORT}}|${KARMADA_AGGREGATED_APISERVER_SECURE_PORT}|g" \
+            | sed "s|{{KARMADA_SEARCH_SECURE_PORT}}|${KARMADA_SEARCH_SECURE_PORT}|g" \
+            | sed "s|{{KARMADA_WEBHOOK_SECURE_PORT}}|${KARMADA_WEBHOOK_SECURE_PORT}|g" \
+            | sed "s|{{KARMADA_SCHEDULER_SECURE_PORT}}|${KARMADA_SCHEDULER_SECURE_PORT}|g" \
+            | sed "s|{{KARMADA_CONTROLLER_MANAGER_SECURE_PORT}}|${KARMADA_CONTROLLER_MANAGER_SECURE_PORT}|g" \
+            > "${LAUNCH_DIR}/${file}"
       fi
     done
   )
@@ -151,42 +154,61 @@ install_kube_artifacts() {
   }
 
   # create namespace for control plane components
-  ${KUBECTL} apply -f "${KARMADA_REPO}/artifacts/deploy/namespace.yaml"
+  ${KUBECTL} apply -f "${ROOT_DIR}/artifacts/namespace.yaml"
 
   # deploy crds
-  TEMP_PATH_CRDS=$(mktemp -d)
-  cp -rf "${KARMADA_REPO}"/charts/karmada/_crds "${TEMP_PATH_CRDS}"
-  fill_caBundle "${KARMADA_REPO}/charts/karmada/_crds/patches/webhook_in_resourcebindings.yaml" \
-              > "${TEMP_PATH_CRDS}/_crds/patches/webhook_in_resourcebindings.yaml"
-  fill_caBundle "${KARMADA_REPO}/charts/karmada/_crds/patches/webhook_in_resourcebindings.yaml" \
-              > "${TEMP_PATH_CRDS}/_crds/patches/webhook_in_clusterresourcebindings.yaml"
-  ${KUBECTL} kustomize "${TEMP_PATH_CRDS}"/_crds | ${KUBECTL} apply -f -
-  rm -rf "${TEMP_PATH_CRDS}"
+  tmp_crds=$(mktemp -d)
+  cp -rf "${ROOT_DIR}"/crds/* "${tmp_crds}"
+  fill_caBundle "${ROOT_DIR}/crds/patches/webhook_in_resourcebindings.yaml" \
+              > "${tmp_crds}/patches/webhook_in_resourcebindings.yaml"
+  fill_caBundle "${ROOT_DIR}/crds/patches/webhook_in_clusterresourcebindings.yaml" \
+              > "${tmp_crds}/patches/webhook_in_clusterresourcebindings.yaml"
+  ${KUBECTL} kustomize "${tmp_crds}" | ${KUBECTL} apply -f -
+  rm -rf "${tmp_crds}"
 
   # deploy webhook configuration
-  fill_caBundle "${KARMADA_REPO}/artifacts/deploy/webhook-configuration.yaml" | ${KUBECTL} apply -f -
+  fill_caBundle "${ROOT_DIR}/artifacts/webhook-configuration.yaml" | ${KUBECTL} apply -f -
 
   # deploy APIService on karmada apiserver for karmada-aggregated-apiserver
-  sed "s/karmada-aggregated-apiserver.karmada-system.svc.cluster.local/localhost/g" \
-        "${KARMADA_REPO}/artifacts/deploy/karmada-aggregated-apiserver-apiservice.yaml" \
+  sed "s/{{KARMADA_AGGREGATED_APISERVER_SECURE_PORT}}/${KARMADA_AGGREGATED_APISERVER_SECURE_PORT}/g" \
+        "${ROOT_DIR}/artifacts/karmada-aggregated-apiserver-apiservice.yaml" \
         | ${KUBECTL} apply -f -
 
   # deploy APIService on karmada apiserver for karmada-search
   # shellcheck disable=SC2002
-  cat "${KARMADA_REPO}/artifacts/deploy/karmada-search-apiservice.yaml" \
-        | sed "s/karmada-search.karmada-system.svc.cluster.local/localhost/g" \
-        | sed '/    namespace: karmada-system/a\
-    port: 6443
-' \
+  sed "s/{{KARMADA_SEARCH_SECURE_PORT}}/${KARMADA_SEARCH_SECURE_PORT}/g" \
+        "${ROOT_DIR}/artifacts/karmada-search-apiservice.yaml" \
         | ${KUBECTL} apply -f -
 
   # deploy cluster proxy rbac for admin
-  ${KUBECTL} apply -f "${KARMADA_REPO}/artifacts/deploy/cluster-proxy-admin-rbac.yaml"
+  ${KUBECTL} apply -f "${ROOT_DIR}/artifacts/cluster-proxy-admin-rbac.yaml"
+}
+
+print_binaries_version() {
+  echo Install karmada with:
+  "${BIN_DIR}"/etcd --version | grep etcd
+  echo "karmada-apiserver version::: $("${BIN_DIR}"/kube-apiserver --version)"
+  echo "kube-controller-manage version::r: $("${BIN_DIR}"/kube-controller-manager --version)"
+  "${BIN_DIR}"/karmada-aggregated-apiserver version
+  "${BIN_DIR}"/karmada-controller-manager version
+  "${BIN_DIR}"/karmada-scheduler version
+  "${BIN_DIR}"/karmada-descheduler version
+  "${BIN_DIR}"/karmada-webhook version
+  "${BIN_DIR}"/karmada-search version
+  #  "${BIN_DIR}"/karmada-scheduler-estimator version
+}
+
+function print_success() {
+  echo "=========================================================="
+  echo "Local Karmada is running."
+  echo -e "\nTo start using your karmada, run:"
+  echo -e "  export KUBECONFIG=${KARMADA_KUBECONFIG}"
 }
 
 clean_dir() {
   echo !!! clean dir
-#  rm -rf "${KARMADA_DIR}"
+  echo clean these dir manually:
+  echo "        rm -rf $KARMADA_DIR $LOG_DIR"
 }
 
 cmd=${1:-help}
